refactor: rename stats module to maths to prepare for character stats
The existing statistical utilities in the stats module are being moved to
a new maths module to avoid naming conflicts with the upcoming character
stats domain concept. This change maintains the same functionality while
creating space for character-specific statistical operations.

Changes:
- Rename src/stats/summary.ts to src/maths/stats.ts
- Create new src/maths/index.ts module entry point
- Update import in src/playbook/sources/pdf.ts
- Move all related test files to test/maths/
