+++ src/maths/index.ts
@@ -1,0 +1,5 @@
+/**
+ * Mathematical utilities and calculations
+ */
+
+export * from "./stats";
