+++ src/maths/stats.ts
