--- src/stats/index.ts
@@ -1,5 +1,0 @@
-/**
- * Statistical utilities and calculations
- */
-
-export * from "./summary";
