+++ test/maths/stats.test.ts
@@ -1,4 +1,4 @@
-import { summary } from "../../src/stats/summary";
+import { summary } from "../../src/maths/stats";
 
 describe("summary", () => {
   describe("edge case: empty input", () => {
